calendar now shows events properly
