+++ src/containers/Calendar/Calendar.jsx
@@ -1,0 +1,64 @@
+import React, { useEffect, useState } from 'react';
+
+import Calendar from 'react-calendar';
+import { getAllEvents } from '../../utils/firebase';
+
+import './Calendar.scss';
+import { calendarDateFormatting, calendarDateGetTime } from '../../utils/helperFunctions';
+
+const CalendarComponent = () => {
+
+    const [month, setMonth] = useState(new Date().getMonth());
+    const [events, setEvents] = useState(null);
+
+    useEffect(() => {
+        const getStuff = async () => {
+            let eventsList = {};
+            (await getAllEvents()).forEach(event =>  {
+                let day = calendarDateFormatting(event?.date?.toDate());
+                eventsList[day] ? eventsList[day].push(event) : eventsList[day] = [event];
+            });
+            setEvents(eventsList);
+        };
+        if(!events) getStuff();
+    });
+
+    const dynamicTileVal = (e) => {
+        return `${(e.date.getDay() === 0 || e.date.getDay() === 6) ? 'weekend' : ''} 
+            ${events?.[calendarDateFormatting(e.date)] ? 'mobileMark' : ''} 
+            ${(e.date.getMonth() !== month) ? 'notMonth' : ''} 
+            calendar__tile`
+    };
+
+    return(
+        <Calendar
+            className = 'calendarPage__calendar'
+            tileClassName = {dynamicTileVal}
+            tileContent = { (e) =>{
+                let day = calendarDateFormatting(e.date);
+                return (
+                    <div className = 'calendar__circle'>
+                        {
+                            //This inserts the correct titles into the calendar boxes
+                            events?.[day] &&
+                            events[day].map((specificEvent, index) => {
+                                return(
+                                    <p key = {index}>
+                                        {`${calendarDateGetTime(specificEvent.date.toDate())} ${specificEvent.name}`}
+                                    </p>
+                                )
+                            })
+                        }
+                    </div>
+                )
+            }
+            }
+            defaultView = {'month'}
+            minDetail = {"month"}
+            onActiveStartDateChange = {(e)=>{setMonth(e.activeStartDate.getMonth())}}
+            locale = {'en-US'}
+
+        />
+    );
+};
+export default CalendarComponent;
