+++ src/containers/Calendar/Calendar.scss
@@ -1,0 +1,7 @@
+@import '../../assets/colors.scss';
+.calendarPage, .calendarPage__particles{
+    height: 100vh;
+}
+.calendarPage{
+    max-width: 100vw;
+}
