+++ src/pages/CalendarPage/CalendarPage.scss
@@ -1,44 +1,78 @@
 @import '../../assets/colors.scss';
-.calendarPage{
-    max-width: 100vw;
-    .calendarPage__calendar{
-        padding: 5%;
-        *>{
+.calendarPage__calendar{
+    padding: 5%;
+    .calendar__tile{
+        display: flex;
+        flex-direction: column;
+        justify-content: space-around;
+        align-items: center;
+        min-height: 3rem;
+        height: 12vh;
+        border: none;
+        .calendar__circle{
+            height: 1vh;
+            width: 1vh;
+            P{
+                display: none;
+            }
         }
+        &::after{
+            content: '';
+            width: 5px;
+            height: 5px;
+            border-radius: 10px;
+        }
+        &.mobileMark::after{
+            background-color: red;
+        }
+        &.weekend{
+            background-color: $gold;
+        }
+        &.notMonth > abbr{
+            color: gray;
+        }
+    }
+    .react-calendar__month-view__weekdays *, .calendar__tile, .react-calendar__navigation *{
+        background-color: $white;
+        border: none;
+        text-decoration: none;
+    }
+    .react-calendar__navigation__arrow{
+        width: 5%;
+    }
+    *{
+        color: $navy;
+        font-family: FuturaBook;
+        font-size: calc(.85vh + .85vw + .425vmin);
+    }
+}
+
+@media only screen and (min-width: 992px){
+    .calendarPage__calendar{
         .calendar__tile{
-            display: flex;
-            flex-direction: column;
-            justify-content: space-around;
-            align-items: center;
-            min-height: 3rem;
-            height: 12vh;
-            border: none;
             .calendar__circle{
-                height: .5rem;
-                width: .5rem;
+                width: 100%;
+                min-height: 5vh;
+                height: 80%;
+                max-width: 95%;
+                background-color: inherit;
                 border-radius: 2rem;
                 justify-self: center;
-                &.active_event{
-                    background-color: red;
+                P{
+                    width: calc(90%);
+                    background-color: inherit;
+                    font-size: 1.7vh;
+                    display: inline-block;
+                    text-align: left;
+                    white-space: nowrap;
+                    overflow: hidden;
+                    text-overflow: ellipsis;
                 }
             }
-            &.weekend{
-                background-color: $gold;
+            
+            &::after{
+                content: none;
             }
-            &.notMonth > abbr{
-                color: gray;
-            }
-        }
-        .react-calendar__month-view__weekdays *, .calendar__tile, .react-calendar__navigation *{
-            background-color: $white;
-            border: none;
-            text-decoration: none;
-
-        }
-        *{
-            color: $navy;
-            font-family: CormorantSemiBold;
-            font-size: calc(.85vh + .85vw + .425vmin);
         }
     }
 }
